Simplify and correct bar logic
Logic is intended to check how many bars to draw by examining 2 bar positions at a time.

Previous version the second conditional was always false (because if barsOn-2*i+1 is not >=0 in the first conditional it certainly won't be >=1 in the second. Instead of >=1, the second conditional needed to be >=-1 in the prior logic.

Simplifying the logic so that the value being checked in the conditional corresponds directly to the number of bars to light solves the problem and makes the flow much more directly understandable.